requirements.txt
captcha>=0.2.1
Pillow>=4.0.0
